example code
This is the master file. You can make change by editing it. If you don't want anyone to review it, you can commit directly to the master branch. If you want someone to review your edits before merging it to the master branch, select the option create a new branch for this commit and start a pull request.